Add files via upload
A heart attack happens when blood supply to the heart becomes significantly reduced or obstructed.The obstruction is typically caused by a buildup of fat, cholesterol, and other chemicals in thearteries leading to the heart (coronary). Plaques are the name given to these cholesterol- and fatty-rich formations. Atherosclerosis describes the process of plaque accumulation.

The offered data includes a table of patient data that includes age, gender, heart rate, systolic bloodpressure, diastolic blood pressure, blood sugar, CK-MB, and troponin values. 100 patients with chestdiscomfort were treated at the hospital and provided with the data. In order to diagnose heart attacks,doctors need the data.